requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.0.2
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
kaleido==0.2.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
panadas==0.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
